experiment with go.work
