+++ go-quark/go.mod
@@ -1,0 +1,3 @@
+module quark
+
+go 1.26.4
