+++ go.work
@@ -1,0 +1,6 @@
+go 1.26.4
+
+use (
+    ./internal/edot
+    ./go-quark
+)
