+++ internal/edot/go.mod
@@ -14,6 +14,7 @@ require (
 	github.com/elastic/beats/v7 v7.0.0-alpha2.0.20260630231048-8d83809f66b2
 	github.com/elastic/elastic-agent v0.0.0-00010101000000-000000000000
 	github.com/elastic/elastic-agent-libs v0.44.0
+	github.com/elastic/go-quark v0.3.0
 	github.com/elastic/mock-es v0.0.0-20250530054253-8c3b6053f9b6
 	github.com/elastic/opentelemetry-collector-components/connector/elasticapmconnector v0.59.0
 	github.com/elastic/opentelemetry-collector-components/connector/profilingmetricsconnector v0.59.0
@@ -110,7 +111,6 @@ require (
 	github.com/elastic/entcollect v0.0.0-20260617022524-00a42d67f0df // indirect
 	github.com/elastic/go-libaudit/v2 v2.6.2 // indirect
 	github.com/elastic/go-licenser v0.4.2 // indirect
-	github.com/elastic/go-quark v0.3.0 // indirect
 	github.com/elastic/opentelemetry-collector-components/internal/elasticattr v0.59.0 // indirect
 	github.com/elastic/opentelemetry-collector-components/internal/sharedcomponent v0.0.0-20250220025958-386ba0c4bced // indirect
 	github.com/elastic/tk-btf v0.2.0 // indirect
